c3

+++ src/app/api/actions/admin-actions.ts
@@ -287,16 +287,21 @@ class AdminActions {
         }
         return makeResponse({ok: true})
     }
+
     static async del_user(data: any) {
-        const {error} = await (await createSuperClient()).auth.admin.deleteUser(data.id)
-        if (error) {
-            return makeResponse({error: (error as any).message})
+        const {data: UData, error: uSError} = await supabaseAdmin.from("users").select("role").eq("id", data.id)
+            .single()
+        if (uSError || !UData) {
+            return makeResponse({error: (uSError as any).message})
         }
-        const {error:dUserError} = await supabaseAdmin.from("users").delete().eq('id',data.id ?? null).select().single()
+        if (UData.role != UserRole.STAFF) {
+            const {error} = await (await createSuperClient()).auth.admin.deleteUser(data.id);
+        }
+        const {error: dUserError} = await supabaseAdmin.from("users").delete().eq('id', data.id ?? null).select().single()
         if (dUserError) {
             return makeResponse({error: (dUserError as any).message})
         }
-        return makeResponse({ok: true})
+        return makeResponse({ok: true, data: UData})
     }
